Analysis/requirements.txt
fastapi
uvicorn[standard]
pandas
matplotlib
python-dateutil
pydantic
pyngrok

# Added for voice service
httpx
python-multipart
# Local speech-to-text
openai-whisper
ffmpeg-python
